requirements.txt
duckdb~=1.3.2
pandas~=2.2.3
pip~=25.2
build~=1.2.2.post1
wheel~=0.45.1
cryptography~=45.0.5
lxml~=6.0.0
tornado~=6.5.1
setuptools~=80.9.0
streamlit~=1.47.1
pywebview~=6.0
requests~=2.32.4